pyproject.toml
[project]
name = "physiboss_models"
version = "0.1.0"
requires-python = ">3.8"
dependencies = [
    "PyGithub",
    "PyYaml"
]
authors = [
  {name = "Vincent Noël", email = "vincent.noel@curie.fr"},
]
description = "Library for using PhysiBoSS-Models"
license = "MIT"
